pyproject.toml
[project]
name = "median-code-backend"
version = "0.1.0"
description = "Median Code Backend - FastAPI Code Generation API"
authors = [
    {name = "Evgeny Aleshin",email = "evgeny@aleshingroup.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.11.9,<3.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)",
    "mako (>=1.3.5,<2.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.32.0,<1.0.0)",
    "sqlalchemy[asyncio] (>=2.0.0,<3.0.0)",
    "asyncpg (>=0.30.0,<1.0.0)",
    "pyjwt[crypto] (>=2.9.0,<3.0.0)",
    "httpx (>=0.28.0,<1.0.0)",
    "python-multipart (>=0.0.20,<1.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)"
]

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-asyncio (>=0.24.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "api_craft", from = "src" },
    { include = "api", from = "src" }
]